Fix settings toggles to save correctly
The Plex Downloads and Plex TV checkboxes in settings were not saving correctly. This is because it was an issue with the booleans being stringified in the db and not being converted back properly. Now they are saved as "true"/"false" in the db and then converted back to booleans in the _load_settings() function. This was the best fix I could come up with, but let me know if you have any better ideas.