Create user/group at runtime in docker-entrypoint.sh
The changes that were made to the Dockerfile didn't quite work since environment variables set at runtime won't change anything that uses them in the Dockerfile. So the user/group creation needs to happen in the docker-entrypoint.sh script so that if the user/group is set as an environment variable it will create a the user/group properly and take ownership of the required folders.